+++ test/Fakes/CleanRegex/Match/Details/Group/TextGroup.php
@@ -27,7 +27,7 @@ class TextGroup implements Group
         throw $this->fail();
     }
 
-    public function textByteLength(): int
+    public function byteLength(): int
     {
         throw $this->fail();
     }