+++ test/Fakes/CleanRegex/Match/Details/Group/ThrowGroup.php
@@ -19,7 +19,7 @@ class ThrowGroup implements Group
         throw $this->fail();
     }
 
-    public function textByteLength(): int
+    public function byteLength(): int
     {
         throw $this->fail();
     }